all: remove atlas config
