+++ cmd_dep_graph.go
@@ -35,7 +35,6 @@ func lbpkr_run_cmd_dep_graph(cmd *commander.Command, args []string) error {
 	var err error
 
 	siteroot := cmd.Flag.Lookup("siteroot").Value.Get().(string)
-	cfgtype := cmd.Flag.Lookup("type").Value.Get().(string)
 	debug := cmd.Flag.Lookup("v").Value.Get().(bool)
 	dotfname := cmd.Flag.Lookup("o").Value.Get().(string)
 	reclvl := cmd.Flag.Lookup("rec-lvl").Value.Get().(int)
@@ -64,7 +63,7 @@ func lbpkr_run_cmd_dep_graph(cmd *commander.Command, args []string) error {
 		)
 	}
 
-	cfg := NewConfig(cfgtype, siteroot)
+	cfg := NewConfig(siteroot)
 	ctx, err := New(cfg, debug)
 	if err != nil {
 		return err
