+++ cmd_deps.go
@@ -28,7 +28,6 @@ ex:
 func lbpkr_run_cmd_deps(cmd *commander.Command, args []string) error {
 	var err error
 
-	cfgtype := cmd.Flag.Lookup("type").Value.Get().(string)
 	debug := cmd.Flag.Lookup("v").Value.Get().(bool)
 	siteroot := cmd.Flag.Lookup("siteroot").Value.Get().(string)
 
@@ -54,7 +53,7 @@ func lbpkr_run_cmd_deps(cmd *commander.Command, args []string) error {
 		)
 	}
 
-	cfg := NewConfig(cfgtype, siteroot)
+	cfg := NewConfig(siteroot)
 	ctx, err := New(cfg, debug)
 	if err != nil {
 		return err
