+++ main.go
@@ -49,5 +49,4 @@ func main() {
 func add_default_options(cmd *commander.Command) {
 	cmd.Flag.String("siteroot", "", "path to site installation")
 	cmd.Flag.Bool("v", false, "enable verbose mode")
-	cmd.Flag.String("type", "lhcb", "config type (lhcb|atlas)")
 }
